Split presentation-contract tests into their own file to stay under the length threshold
SettingsWindowNavigationRoutingTests.swift reached 540 lines (untracked
files must stay under 500). The Dock-deminiaturize, failed-activation, and
re-entrant teardown tests move to SettingsWindowPresentationContractTests
(wired into the pbxproj, nested under the same serialized parent suite);
the window helpers become target-internal and are shared.
